Add an icon legend to a reactable table

Use `add_icon_legend()` to place a legend either below or above a reactable or reactablefmtr table. The legend can be used to display the icon set used within `icon_sets()`. The legend can be aligned to either the right, left, or center of the table. Custom labels can be applied to the upper and lower bounds of the legend.

add_icon_legend(
  table = NULL,
  icon_set = NULL,
  show_labels = TRUE,
  labels = NULL,
  position = "below",
  align = "right",
  margin = NULL
)

Arguments

table
A reactable table.

icon_set
The icon set to be displayed in the legend. Options are "ski rating", "medals", and "batteries". Default is NULL.

show_labels
Logical. Show or hide the labels next to the legend. Default is TRUE.

labels
Assign label to each icon in the specified icon set. Number of labels must match the number of icons in the icon set. Default is NULL.

position
The position of the legend in relation to the table. Options are 'above' or 'below'. Default is 'below'.

align
The horizontal alignment of the legend. Options are 'left', 'right', or 'center'. Default is 'right'.

margin
Use margin() to set the margin around the legend (top, right, bottom, left). Default is NULL.

Value

a function that adds a source below a reactable table.

Examples

if (FALSE) {
## Create the reactable table and then pipe in the legend
library(dplyr)
data <- iris[10:29, ]
table <- reactable(data,
columns = list(Petal.Length = colDef(
cell = icon_sets(data, icon_set = "medals"))))

table %>%
  add_icon_legend(icon_set = "medals")

## The legend can be placed below or above the table
## and aligned either to the left, right, or center
table %>%
  add_icon_legend(icon_set = "medals", position = "above", align = "left")

## Add custom labels to each icon in the legend
table %>%
  add_icon_legend(icon_set = "medals", labels = c("Shortest Length","Avg Length","Longest Length"))
}