Add color scales to cells in a column

The `color_scales()` function conditionally colors each cell of a column depending on their value in relation to other values in that particular column. The colors can be provided within a vector in `colors` or via another column in the dataset by referencing the column by name with `color_ref`. The opacity of the colors provided can be adjusted by providing a value between 0 and 1 in `opacity`. `text_color` can be used to change the color of the values. If values are displayed within a dark-colored background, `brighten_text` will display the values in white text so they are more visible. The color of `brighten_text_color` can be changed to a color other than white if desired. If the user wants to assign colors row-wise instead of column-wise, set `span` equal to TRUE to apply across all columns. Or can provide the names of the columns by either column name or column position number to apply to only a subset of the columns. `color_scales()` should be placed within the style argument in reactable::colDef.

color_scales(
  data,
  colors = c("#67a9cf", "#f8fcf8", "#ef8a62"),
  color_ref = NULL,
  opacity = 1,
  text_color = "black",
  show_text = TRUE,
  brighten_text = TRUE,
  brighten_text_color = "white",
  bold_text = FALSE,
  span = FALSE,
  animation = "1s ease"
)

Arguments

data
Dataset containing at least one numeric column.

colors
A vector of colors to color the cells. Colors should be given in order from low values to high values. Default colors provided are blue-white-orange: c("#67a9cf", "#f8fcf8", "#ef8a62"). Can use R's built-in colors or other color packages.

color_ref
Optionally assign colors to from another column by providing the name of the column containing the colors in quotes. Only one color can be provided per row. Default is NULL.

opacity
A value between 0 and 1 that adjusts the opacity in colors. A value of 0 is fully transparent, a value of 1 is fully opaque. Default is 1.

text_color
Assigns text color to values. Default is black.

show_text
Logical: show text or hide text. Default is TRUE.

brighten_text
Logical: automatically assign color to text based on background color of cell. Text within dark-colored backgrounds will turn white, text within light-colored backgrounds will be black. Default is TRUE.

brighten_text_color
Assigns text color to values if values are within a dark-colored backgrounds. Default is white.

bold_text
Logical: bold text. Default is FALSE.

span
Optionally apply colors to values across multiple columns instead of by each column. To apply across all columns set to TRUE. If applying to a set of columns, can provide either column names or column positions. Default is set to FALSE.

animation
Control the duration and timing function of the animation when sorting/updating values shown on a page. See [CSS transitions](https://developer.mozilla.org/en-US/docs/Web/CSS/transition) for available timing functions and examples. Animation can be turned off by setting to "none". Default is "1s ease".

Value

a function that applies conditional colors to a column of numeric values.

Examples

data <- iris[10:29, ]

## By default, the colors_scales() function uses a blue-white-orange three-color pattern
reactable(data,
 columns = list(
 Petal.Length = colDef(style = color_scales(data))))

## If only two colors are desired,
## you can specify them with colors = 'c(color1, color2)';
reactable(data,
 columns = list(
 Petal.Length = colDef(style = color_scales(data,
 colors = c("red", "green")))))

## Apply color_scales() across all numeric columns using reactable::defaultColDef
reactable(data,
defaultColDef = colDef(style = color_scales(data)))

## Use span to apply colors to values in relation to the entire dataset
reactable(data,
defaultColDef = colDef(style = color_scales(data, span = TRUE)))

## Span can take column names
reactable(data,
defaultColDef = colDef(style = color_scales(data, span = c("Sepal.Length", "Sepal.Width"))))

## Or it can also take column positions instead
reactable(data,
defaultColDef = colDef(style = color_scales(data, span = 1:2)))