Highlights the minimum and maximum value in a column

The `highlight_min_max()` function assigns a font color and/or background color to both the minimum and maximum values in a column. It should be placed within the style argument in reactable::colDef.

highlight_min_max(
  data,
  min_font_color = "black",
  max_font_color = "black",
  min_highlighter = NULL,
  max_highlighter = NULL
)

Arguments

data
Dataset containing at least one numeric column.

min_font_color
color to assign to minimum value in a column. Default color is black.

max_font_color
color to assign to maximum value in a column. Default color is black.

min_highlighter
color to assign the background of a cell containing minimum value in a column.

max_highlighter
color to assign the background of a cell containing maximum value in a column.

Examples

library(reactable)
data <- MASS::road[11:17, ]

## By default, the minimum and maximum values are bold with a black font color
reactable(data, defaultColDef = colDef(
    style = highlight_min_max(data)))

## Assign a different font color to the min and max values
reactable(data, defaultColDef = colDef(
    style = highlight_min_max(data, min_font_color = "red", max_font_color = "green")))

## Highlight the background of the cell for the min and max values in each column
reactable(data, defaultColDef = colDef(
    style = highlight_min_max(data, min_highlighter = "salmon", max_highlighter = "skyblue")))